% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_api_keys.R
\name{check_api_keys}
\alias{check_api_keys}
\title{Checks if your API creds and base url are set as environment vars}
\usage{
check_api_keys()
}
\value{
bool
}
\description{
Checks if your API creds and base url are set as environment vars
}
