% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtricks.R
\name{checkRenviron}
\alias{checkRenviron}
\title{INTERNAL: Check for requisite .Renviron variables; alert if missing}
\usage{
checkRenviron(var)
}
\description{
INTERNAL: Check for requisite .Renviron variables; alert if missing
}
\keyword{internal}
