% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_contact.R
\name{create_contact}
\alias{create_contact}
\title{Create a new contact in a mailing list}
\usage{
create_contact(mailinglistID, body)
}
\arguments{
\item{mailinglistID}{check \code{qualtRics::all_mailinglists()}}

\item{body}{named list (careful with names, e.g. externalDataRef -> compare api documentation)}
}
\value{
content of request as returned by \code{httr::content()}
}
\description{
Create a new contact in a mailing list
}
\examples{
\dontrun{
mailinglistID <- qualtRics::all_mailinglists()$id[1]
body <- list(firstName = "hello", lastName = "world", email = "hello@world.ch", externalDataRef = "E", language = "de", unsubscribed = FALSE)
create_contact(mailinglistID = mailinglistID, body = body)
}
}
\seealso{
\url{https://api.qualtrics.com/api-reference/b3A6NjEwMzA-create-contact}
}
