% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-thoscha.R
\name{download_translations}
\alias{download_translations}
\title{Download translations for multi-lingual languages}
\usage{
download_translations(survey_id, languages = c("DE", "EN"))
}
\arguments{
\item{languages}{char vector of all language codes used in your survey, defaults to \code{c("DE", "EN")}}
}
\description{
Please install qualtricks from ivt-baug-ethz and set up your Qualtrics credentials.
}
\author{
Thomas Schatzmann \email\link{thomas.schatzmann@ivt.baug.ethz.ch}
}
