% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Email.R
\name{Email}
\alias{Email}
\title{R6 Class representing a single E-Mail}
\description{
R6 Class representing a single E-Mail

R6 Class representing a single E-Mail
}
\details{
It let's you define the recipient, the email itself and the connection and is
a helper object for working with the \code{blastula} package.
}
\examples{
\dontrun{
qr <- add_image(file = "/Users/schnickdani/Downloads/SV_6u4YzY24OAfhWfQ-qrcode.png", width = 300)
logo <- add_image(file = "/Users/schnickdani/Downloads/eth-logo.png", width = 150, align = "left")
firstname <- "Daniel"
lastname <- "Heimgartner"

body <-
  htmltools::HTML(
    glue::glue("<h1>Sehr geehrte(r) {firstname} {lastname}</h1>",
               "<p>Vielen Dank für Ihre Teilnahme an unserer Studie zum Thema <b>Homeoffice</b>.</p>",
               "<p>Sie gelangen mittels QR-Code zur Umfrage.</p>",
               "{qr}",
               "<p>Freundliche Grüsse<br>IVT Projektteam<br>Daniel Heimgartner<br>{logo}</p>"))

footer <- glue::glue("Sent by Daniel Heimgartner on {Sys.Date()}")

em <- Email$new()
em
em$write_mail(body = body, footer = footer)
em
em$send_mail()
em$mail_adress <- "foo@gmail.com"
em$send_mail(subject = "Testing")
em
}

}
\seealso{
\code{blastula} package

\code{blastula::compose_email}

\code{blastula::smtp_send}
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{mail_adress}}{Mail adress of the recipient}

\item{\code{user}}{The username for the email account. Typically, this is the
email address associated with the account. In the case of ETH this is
your user name (e.g. daniehei);
Reference: \code{blastula} package}

\item{\code{sender}}{The adress of the sender. In the case of ETH this is your
mail adress (e.g. \href{mailto:daniel.heimgartner@ivt.ethz.ch}{daniel.heimgartner@ivt.ethz.ch})}

\item{\code{provider}}{An optional email provider shortname for autocompleting
SMTP configuration details (the \code{host}, \code{port}, \code{use_ssl} options). Options
currently include gmail, outlook, and office365. If nothing is provided
then values for \code{host}, \code{port}, and \code{use_ssl} are expected;
Reference: \code{blastula} package}

\item{\code{host}}{Configuration info for the SMTP server. The \code{host} and \code{port}
parameters are the address and port for the SMTP server;
Reference: \code{blastula} package}

\item{\code{port}}{Configuration info for the SMTP server. The \code{host} and \code{port}
parameters are the address and port for the SMTP server;
Reference: \code{blastula} package}

\item{\code{use_ssl}}{is an option as to whether to use SSL: supply a \code{TRUE} or \code{FALSE} value.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Email$new()}}
\item \href{#method-write_mail}{\code{Email$write_mail()}}
\item \href{#method-send_mail}{\code{Email$send_mail()}}
\item \href{#method-print}{\code{Email$print()}}
\item \href{#method-clone}{\code{Email$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Create a new email object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Email$new(
  mail_adress = NULL,
  user = "daniehei",
  sender = "daniel.heimgartner@ivt.baug.ethz.ch",
  provider = "outlook",
  host = "mail.ethz.ch",
  port = 587,
  use_ssl = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{mail_adress}}{Mail adress of recipient.}

\item{\code{user}}{User name for email account.}

\item{\code{sender}}{The adress of the sender.}

\item{\code{provider}}{An email provider shortname (gmail, outlook or office 365).}

\item{\code{use_ssl}}{Use SSl.}

\item{\code{Configuration}}{info for the SMTP server.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A new \code{Email} object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-write_mail"></a>}}
\if{latex}{\out{\hypertarget{method-write_mail}{}}}
\subsection{Method \code{write_mail()}}{
Write an email. Wrapper around \code{blastula::compose_email}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Email$write_mail(...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{...}}{arguments passed to \code{blastula::compose_email}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-send_mail"></a>}}
\if{latex}{\out{\hypertarget{method-send_mail}{}}}
\subsection{Method \code{send_mail()}}{
Send the email. Wrapper around \code{blastula::smtp_send}
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Email$send_mail(subject, ...)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{subject}}{Subject for the email}

\item{\code{...}}{args passed to \code{blastula::smtp_send}}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print method for Email. Renders the email in the Viewer.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Email$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Email$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
