% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previewIntercept.R
\name{formatInterceptBookmarklet}
\alias{formatInterceptBookmarklet}
\title{Format JavaScript bookmarklet that triggers Qualtrics Intercept Edit Preview}
\usage{
formatInterceptBookmarklet(interceptId, zoneURL = Sys.getenv("QSI_ZONE_URL"))
}
\arguments{
\item{interceptId}{string}

\item{zoneURL}{string}
}
\value{
character
}
\description{
Format JavaScript bookmarklet that triggers Qualtrics Intercept Edit Preview
}
