% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-thoscha.R
\name{generate_codebook}
\alias{generate_codebook}
\title{Generates a code book for your survey}
\usage{
generate_codebook(survey_id, languages = c("DE", "EN"))
}
\arguments{
\item{languages}{char vector of all language codes used in your survey, defaults to \code{c("DE", "EN")}
(also passed to \code{download_translations})}
}
\value{

}
\description{
Calls download_translations internally
}
\examples{
\dontrun{
  survey_id <- "SV_6u4YzY24OAfhWfQ"
  generate_codebook(survey_id)
}
}
\author{
Thomas Schatzmann \email\link{thomas.schatzmann@ivt.baug.ethz.ch}
}
