% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_codes.R
\name{generate_codes}
\alias{generate_codes}
\title{Generate access codes}
\usage{
generate_codes(n, length = 4, write = FALSE)
}
\arguments{
\item{n}{number of codes to be generated}

\item{length}{length of codes and passwords}

\item{write}{logical, write to yaml file}
}
\value{
data.frame
}
\description{
Generate access codes
}
