% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codebook.R
\name{get_codebook}
\alias{get_codebook}
\title{Generate a codebook}
\usage{
get_codebook(
  survey_id,
  spec = NULL,
  encoding = "utf-8",
  add_translations = TRUE,
  ...
)
}
\arguments{
\item{spec}{optionally passed to \code{tibblify()}, defaults to \code{NULL}}

\item{encoding}{passed to \code{get_questions}, defaults to \code{"utf-8"}}

\item{add_translations}{should translations data.frame be added as attribute, defaults to \code{TRUE}
Retrieve via \code{attributes(codebook, "translations")}}

\item{...}{not implemented}
}
\value{

}
\description{
Generate a codebook
}
\seealso{
\code{generate_codebook()}
}
