% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_label.R
\name{get_label}
\alias{get_label}
\title{Get labels from data.frame}
\usage{
get_label(df, ...)
}
\arguments{
\item{df}{data.frame}

\item{...}{passed to \code{Hmisc::label()}}
}
\description{
Usually the df is retrieved via \code{qualtRics::fetch_survey()}
}
\examples{
\dontrun{qualtRics::all_surveys()$id[4] \%>\% qualtRics::fetch_survey() \%>\% get_label()}
}
