% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/organizations.R
\name{getOrganization}
\alias{getOrganization}
\title{Get Organization}
\usage{
getOrganization(
  yourdatacenterid = Sys.getenv("QUALTRICS_DATACENTERID"),
  yourapitoken = Sys.getenv("QSI_TOKEN"),
  organizationId
)
}
\arguments{
\item{yourdatacenterid}{Set in .Renviron}

\item{yourapitoken}{Set in .Renviron}

\item{organizationId}{Set in .Renviron or elsewhere}
}
\value{
data.frame
}
\description{
Gets general information about an organization
https://api.qualtrics.com/reference#get-brand-info
}
\examples{
\dontrun{
getOrganization(organizationId="hbp")
}
}
