% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtricsHelpers.R
\name{getProjectStats}
\alias{getProjectStats}
\title{Wrapper for getCreativeStats and getInterceptStats API calls}
\usage{
getProjectStats(user = Sys.getenv("QSI_USER"),
  token = Sys.getenv("QSI_TOKEN"), format = "JSON", projectId, startDate,
  endDate, timeZone = "America\%2FNew_York", dataType, interval)
}
\arguments{
\item{user}{}

\item{token}{}

\item{format}{}

\item{projectId}{}

\item{startDate}{}

\item{endDate}{}

\item{timeZone}{}

\item{dataType}{}

\item{interval}{}
}
\description{
Wrapper for getCreativeStats and getInterceptStats API calls
}
