% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\name{listAllSurveys}
\alias{listAllSurveys}
\title{List All Surveys}
\usage{
listAllSurveys(...)
}
\arguments{
\item{...}{args for listSurveys}
}
\value{

}
\description{
like listSurveys but does the pagination
}
\examples{
listAllSurveys()
}
