% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtricks.R
\name{listCreatives}
\alias{listCreatives}
\title{Scrape Qualtrics Site Intercept creative ids and names for use in API call}
\usage{
listCreatives(creativeHomeURL = Sys.getenv("CREATIVE_HOME_URL"),
  cssSelector = "tr[creatives='creatives']",
  script = paste0(readLines(paste0(path.package("qualtricks"),
  "/scrape_qsi_creative.js"), warn = FALSE), collapse = "\\n"))
}
\arguments{
\item{creativeHomeURL}{URL for creative project list on legacy UI; defaults to .Renviron}

\item{cssSelector}{CSS Selector matching an individual creative}

\item{script}{Path to text file containing JavaScript to execute on page to wrangle the creatives into an array}
}
\value{
data.frame
}
\description{
Uses Experience Management UI
}
\examples{
listCreatives()
}
