% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtricks.R
\name{listIntercepts}
\alias{listIntercepts}
\title{Scrape Qualtrics Site Intercept intercept ids and names for use in API call}
\usage{
listIntercepts(interceptHomeURL = Sys.getenv("INTERCEPT_HOME_URL"),
  interceptZoneId = Sys.getenv("INTERCEPT_ZONE_ID"))
}
\arguments{
\item{interceptHomeURL}{defaults to .Renviron}

\item{interceptZoneId}{defaults to .Renviron}
}
\value{
data.frame
}
\description{
Uses legacy UI (not Experience Management)
Currently less flexible than listCreatives() since more selectors are needed for zone nav and they're still hardcoded
}
\examples{
listIntercepts()
}
