% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_intercept.R
\name{listInterceptsAndCreatives}
\alias{listInterceptsAndCreatives}
\title{List Intercepts and Creatives}
\usage{
listInterceptsAndCreatives(
  summaryReportURL = Sys.getenv("QUALTRICS_SUMMARY_REPORT_URL"),
 
    pageScript = "var c = [];\\n                                   document.querySelectorAll(\\"table.StatsTable tbody tr td:nth-of-type(1) a\\").forEach(function(x) {c.push(x.getAttribute(\\"clickcallback\\"))})\\n                                   return c"
)
}
\arguments{
\item{summaryReportURL}{String. URL for license summary report}

\item{pageScript}{String. Javascript to scrape the tables}
}
\value{
data.frame
}
\description{
Scrape intercept and creative ids and names from admin Summary Report page for use in API calls\cr
Uses RSelenium to open and scrape Chrome
}
\examples{
\dontrun{
listInterceptsAndCreatives()
}
}
