% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtricsHelpers.R
\name{listProjects}
\alias{listProjects}
\title{Scrape Qualtrics Creative and Intercept list for ids to use in API calls}
\usage{
listProjects()
}
\description{
Scrape Qualtrics Creative and Intercept list for ids to use in API calls
}
