% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surveys.R
\name{listSurveys}
\alias{listSurveys}
\title{List Surveys
https://api.qualtrics.com/guides/reference/surveys.json/paths/~1surveys/get}
\usage{
listSurveys(
  yourdatacenterid = Sys.getenv("QUALTRICS_DATACENTERID"),
  yourapitoken = Sys.getenv("QSI_TOKEN"),
  offset = 0,
  responseOnly = FALSE
)
}
\arguments{
\item{yourdatacenterid}{Set in .Renviron}

\item{yourapitoken}{Set in .Renviron}

\item{offset}{paging, defaults to 0 (first page only)}

\item{responseOnly}{Return response() object only? Defaults to FALSE.}
}
\value{
Response or data.frame
}
\description{
List Surveys
https://api.qualtrics.com/guides/reference/surveys.json/paths/~1surveys/get
}
