% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/location2state.R
\name{ll2county}
\alias{ll2county}
\title{Latitude/longitude to county}
\usage{
ll2county(x)
}
\arguments{
\item{x}{Input data frame with lat and lng coordinate columns}
}
\value{
Names of counties in which coordinates are located
}
\description{
Converts geographical coordinates to US counties.
}
