% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{name_parser}
\alias{name_parser}
\title{Make ugly names less ugly}
\usage{
name_parser(x)
}
\arguments{
\item{x}{character vector}
}
\description{
Usually applied to survey or col_map or something the like...
}
