% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qualtricks.R
\name{qualtricsLogin}
\alias{qualtricsLogin}
\title{INTERNAL: Log in to Qualtrics from institutional URL}
\usage{
qualtricsLogin()
}
\description{
INTERNAL: Log in to Qualtrics from institutional URL
}
\keyword{internal}
