% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qualtrics.R
\name{read_qualtrics}
\alias{read_qualtrics}
\title{read qualtrics CSV}
\usage{
read_qualtrics(x)
}
\arguments{
\item{Path}{to CSV exported from Qualtrics}
}
\value{
A tibble
}
\description{
Reads and formats data exported from qualtrics.
}
