% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_qualtrics.R
\name{recode_likert}
\alias{recode_likert}
\title{Recode likert items}
\usage{
recode_likert(x)
}
\arguments{
\item{x}{Input data frame with likert columns}
}
\value{
A data frame with likert choice columns replaced with integer values
}
\description{
Converts likert columns into integer columns
}
