% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_translation_labels.R
\name{set_translation_labels}
\alias{set_translation_labels}
\title{Set translation labels}
\usage{
set_translation_labels(
  survey,
  codebook,
  .exclude = rlang::quos(QuestionType == "TE" | stringr::str_detect(DataExportTag,
    "^Q"))
)
}
\arguments{
\item{survey}{data.frame}

\item{codebook}{see \code{get_codebook}}

\item{.exclude}{passed to filter... better stick with the defaults... otherwise look into function body to understand what happens :S}
}
\value{
returns labels invisible and sets \code{labelr::labels}
}
\description{
Has to be called before \code{translate_survey} (has sideffects -> sets labels \code{labelr::labels$get()})
}
