% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{strip_html}
\alias{strip_html}
\title{Strips html tags}
\usage{
strip_html(x)
}
\arguments{
\item{x}{a string with html tags}
}
\value{

}
\description{
Strips html tags
}
\references{
Taken from \code{qualtRics:::remove_html}
}
