% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_answer.R
\name{translate_answer}
\alias{translate_answer}
\title{Translate one answer vector}
\usage{
translate_answer(
  v,
  from = c("Answer", "AnswerValue"),
  codebook = survey$codebook
)
}
\arguments{
\item{v}{answer vector unquoted expression (usually a column from the survey data.frame)}

\item{from}{hint what v contains (if numeric -> AnswerValue, if main language to foreign -> Answer)}

\item{codebook}{see \code{get_codebook}}
}
\description{
Translate one answer vector
}
