% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/translate_survey.R
\name{translate_survey}
\alias{translate_survey}
\title{Translate the survey}
\usage{
translate_survey(survey, type = c("AnswerTranslation", "AnswerValue"), ...)
}
\arguments{
\item{survey}{survey data.frame}

\item{type}{which labels should be used?}

\item{...}{}
}
\description{
Translate the survey
}
