% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_contact.R
\name{update_contact}
\alias{update_contact}
\title{Update a contact in your mailing list}
\usage{
update_contact(mailinglistID, contactID, body)
}
\arguments{
\item{mailinglistID}{check \code{qualtRics::all_mailinglists()}}

\item{contactID}{check \code{qualtRics::fetch_mailinglist()}}

\item{body}{named list (careful with names, e.g. externalDataReference -> does not match api documenation... internal strangeness)}
}
\value{
content of request as returned by \code{httr::content()}
}
\description{
Update a contact in your mailing list
}
\examples{
\dontrun{
mailinglist <- qualtRics::all_mailinglists()
mailinglistID <- mailinglist$id[1]
m <- fetch_mailinglist(mailinglistID)
contactID <- m$id[1]
code <- generate_codes(n = 1)$code
body <- list(firstName = "hello", lastName = "world", externalDataReference = code)
update_contact(mailinglistID = mailinglistID, contactID = contactID, body = body)
fetch_mailinglist(mailinglistID)
}
}
\seealso{
\url{https://api.qualtrics.com/api-reference/b3A6NjEwMzA-create-contact}
}
