% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/legacy_crud.R
\name{updateSurvey}
\alias{updateSurvey}
\title{Update Survey}
\usage{
updateSurvey(
  surveyId,
  name = NULL,
  isActive = c(TRUE, FALSE),
  expirationStartDate = NULL,
  expirationEndDate = NULL,
  ownerID = NULL,
  yourapitoken = Sys.getenv("QSI_TOKEN")
)
}
\arguments{
\item{surveyId}{String. Survey ID}

\item{name}{String. New survey name}

\item{isActive}{Boolean. Survey status}

\item{expirationStartDate}{String. Defines the active time range for the survey. Example: {"startDate":"2016-01-01T01:00:00Z", "endDate":"2016-03-01T01:00:00Z"}. See Dates and Times for more information on the date and time format.}

\item{expirationEndDate}{String. See expirationStartDate}

\item{ownerID}{String. The new owner of the survey. Note that the caller must own the survey to set a new owner.}

\item{yourapitoken}{String from renviron}
}
\value{
response() object
}
\description{
Updates a survey \cr
https://api.qualtrics.com/reference#update-survey
[] not all args handled atm
}
\examples{
\dontrun{
library(qualtRics)
to_deactivate <- all_surveys() \%>\%
  mutate(lastMod = as.Date(lastModified)) \%>\%
  filter(isActive == TRUE & lastMod <= "2017-01-01")

updateSurvey(to_deactivate$id[1], isActive=FALSE)
}
}
