% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/site_intercept.R
\name{getProjectStats}
\alias{getProjectStats}
\title{Wrapper for getCreativeStats and getInterceptStats API calls}
\usage{
getProjectStats(
  projectId,
  startDate,
  endDate,
  timeZone = "America\%2FNew_York",
  dataType = c("Impressions, Clicks"),
  interval = c("Hour", "Day", "Month")
)
}
\arguments{
\item{projectId}{Intercept ID or Creative ID; appropriate call will be determined from this string}

\item{startDate}{YYYY-MM-DD}

\item{endDate}{YYYY-MM-DD}

\item{timeZone}{The timezone to aggregate responses relative to. http://php.net/manual/en/timezones.php}

\item{dataType}{Impressions, Clicks}

\item{interval}{Hour (default), Day, Month}
}
\value{
data.frame
}
\description{
https://survey.qualtrics.com/WRAPI/SiteIntercept/docs.php#getInterceptStats_2.4
}
\examples{
\dontrun{
projects <- listInterceptsAndCreatives()
getProjectStats(projectId=projects$id[1], Sys.Date()-365, Sys.Date(), dataType = "Impressions", interval="Month")
}
}
