% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/users.R
\name{listUsers}
\alias{listUsers}
\title{List Users
https://api.qualtrics.com/reference/list-users}
\usage{
listUsers(yourdatacenterid = Sys.getenv("QUALTRICS_DATACENTERID"),
  yourapitoken = Sys.getenv("QSI_TOKEN"), responseOnly = FALSE)
}
\arguments{
\item{yourdatacenterid}{Set in .Renviron}

\item{yourapitoken}{Set in .Renviron}

\item{responseOnly}{Return response() object only? Defaults to FALSE.}
}
\value{
Response
}
\description{
List Users
https://api.qualtrics.com/reference/list-users
}
\examples{
\dontrun{
listUsers()
}
}
