% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/previewIntercept.R
\name{startSelServer}
\alias{startSelServer}
\title{Start a selenium server and browser with or without Docker}
\usage{
startSelServer(browser = "chrome", useDocker = FALSE)
}
\arguments{
\item{browser}{defaults to chrome}

\item{useDocker}{probably don't need this}
}
\value{
remDr
}
\description{
Start a selenium server and browser with or without Docker
}
